% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_boxplot.R
\name{create_boxplot}
\alias{create_boxplot}
\title{Boxplot Function}
\usage{
create_boxplot(data = DRG_data, payment_type = "Average Medicare Payments")
}
\arguments{
\item{data}{the data frame "DRG_data"}

\item{payment_type}{a string name indicating the payment type}
}
\value{
\code{create_boxplot} returns a boxplot; \code{Medipay_stat} returns a sentence containing requested statistic.
}
\description{
\code{create_boxplot} makes a boxplot of payments by DRG code;
}
\examples{
create_boxplot(DRG_data, payment_type = "Average Medicare Payments")
}
