% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Medipay_stat.R
\name{Medipay_stat}
\alias{Medipay_stat}
\title{Calculation Function}
\usage{
Medipay_stat(x)
}
\arguments{
\item{x}{a string name indicating the statistic to be calculated}
}
\value{
\code{Medipay_stat} returns a sentence containing requested statistic.
}
\description{
\code{Medipay_stat} calculates statistics over all of the DRG codes for average Medicare payments.
}
\examples{
Medipay_stat("mean")
}
